refactor: rewrite boards and columns in ts
